Create CustomCssComponent
